refactor: 향후 api 구조 변경을 대비하기 위해 v1 prefix 추가
- controller, dto에 v1 디렉토리 구조 추가
- api url에 api/vl preifx 추가

#9
