+++ backend/check-tracker/src/main/java/com/hohohehe/checktracker/dto/v1/request/CheckListRequest.java
@@ -1,4 +1,4 @@
-package com.hohohehe.checktracker.dto.request;
+package com.hohohehe.checktracker.dto.v1.request;
 
 public record CheckListRequest(
         String title
