+++ backend/check-tracker/src/test/java/com/hohohehe/checktracker/contoller/v1/CheckListControllerTest.java
@@ -1,9 +1,9 @@
-package com.hohohehe.checktracker.contoller;
+package com.hohohehe.checktracker.contoller.v1;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hohohehe.checktracker.config.MockServiceConfig;
 import com.hohohehe.checktracker.domain.CheckList;
-import com.hohohehe.checktracker.dto.request.CheckListRequest;
+import com.hohohehe.checktracker.dto.v1.request.CheckListRequest;
 import com.hohohehe.checktracker.service.CheckListService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -48,7 +48,7 @@ class CheckListControllerTest {
         given(checkListService.searchCheckList()).willReturn(createCheckList(checkListId));
 
         // when & then
-        mvc.perform(get("/checkLists"))
+        mvc.perform(get("/api/v1/checklists"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                 .andExpect(jsonPath("$[0].checkListId").value(checkListId))
@@ -63,7 +63,7 @@ class CheckListControllerTest {
         given(checkListService.saveCheckList(any(CheckList.class))).willReturn(any(CheckList.class));
 
         // when & then
-        mvc.perform(post("/checkLists")
+        mvc.perform(post("/api/v1/checklists")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(mapper.writeValueAsString(request)))
                 .andExpect(status().isOk())
